moved cf_units to __init__
